requirements.txt
tensorflow-gpu==2.10.0
keras==2.10.0
scikit-learn==1.3.2
opencv-python
matplotlib
pandas
numpy
tqdm